[no-verify] Update `eval` to use new east refactor.
